pyproject.toml
[project]
name = "marimaker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "backgroundremover==0.1.9",
    # ML stack — MUST be pinned
    "numpy==1.21.5",
    "scikit-image==0.19.1",
    "scipy==1.7.3",
    "pymatting==1.1.5",
    "pywavelets==1.2.0",
    "numba==0.55.0",
    "llvmlite==0.38.0",
    # App stack
    "flask==2.0.2",
    "werkzeug==2.0.2",
    "moviepy==1.0.3",
    "ffmpeg-python==0.2.0",
    "pillow==8.1.1",
    "requests==2.27.1",
    "gdown==4.2.0",
    "gunicorn==20.1.0",
    "waitress==2.0.0",
    "torch==1.10.1",
    "torchvision==0.11.2",
]
